fix non-numeric value warning in tests when using php 7

+++ tests/AlexaRequestValidatorTest.php
@@ -47,7 +47,7 @@ class AlexaRequestValidatorTest extends TestCase
      */
     public function it_returns_true_when_the_request_in_inside_the_timestamp_tolerance()
     {
-        $date = date('Y-m-d\TH:i:s\Z') - 110;
+        $date = (int)date('Y-m-d\TH:i:s\Z') - 110;
         $request = ['request' => ['timestamp' => $date]];
         $validator = new AlexaRequestValidator('foo', json_encode($request), 'https://s3.amazonaws.com/echo.api/echo-api-cert.pem', 'foo');
         $this->assertTrue($validator->requestHasNotTimedOut());
@@ -59,7 +59,7 @@ class AlexaRequestValidatorTest extends TestCase
     public function it_throws_an_exception_when_the_request_has_timed_out()
     {
         // Check the maximum amount of time allowed by amazon, by default this is set to 120 seconds so anything after that should fail
-        $date = date('Y-m-d\TH:i:s\Z') - 150;
+        $date = (int)date('Y-m-d\TH:i:s\Z') - 150;
         $request = ['request' => ['timestamp' => $date]];
         $this->expectException(AlexaValidationException::class);
         $validator = new AlexaRequestValidator('foo', json_encode($request), 'https://s3.amazonaws.com/echo.api/echo-api-cert.pem', 'foo');
